fix: rtpc v03 object id stuff
